solution/requirements.txt
bitstring==3.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
cryptography==3.4.7
defusedxml==0.6.0
elementpath==2.2.2
idna==2.10
importlib-resources==5.1.3
isodate==0.6.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.0
more-itertools==8.7.0
portend==2.7.1
pycparser==2.20
pyOpenSSL==20.0.1
pysaml2==6.5.1
python-dateutil==2.8.1
python3-saml==1.10.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tempora==4.0.2
urllib3==1.26.4
xmlschema==1.6.2
xmlsec==1.3.10
zc.lockfile==2.0
